index keywords also with an "XK" prefix
That is, also index the keyword strings as they originally appear,
without any normalization or stemming or splitting into words.

Use those index entries to implement `notdeft-xapian-list-all-keywords'
as a better replacement for `notdeft-xapian-list-all-keyword-words'.

The new function is able to list keywords such as "@home" or
"los-angeles" without dropped characters or split words, despite the
Xapian term generator not treating '@' or '-' as word characters.