also refer to `notdeft-xapian-program' when building the program
That is, if `notdeft-xapian-program' is already set then just have
`notdeft-xapian-make' build the binary there.

This even if set to a default value by `notdeft-xapian', which can
only happen if an executable already exists at the location, in which
case better build an updated binary there to avoid old potentially
incompatible copies left over to be used accidentally.