add `notdeft-xapian-order-by' customization option
Add it as an option to specify search result ordering. It is a
replacement for `notdeft-xapian-order-by-time', and so this is a
breaking change for anyone who had customized that variable.

This is a more general mechanism allowing other orderings to also be
specified using the same parameter, including ordering by non-directory
file `name', which corresponds to the existing "!file" query directive.

Also change search function signatures to take a plist of keyword
arguments to allow an :order-by parameter to be passed around, among
others. Have `notdeft-search-find-file-function' assume that its
implementations have that kind of signature.

For example adapt `notdeft-lucky-find-file' signature accordingly, but
do not change its interactive use.

Also fix brokenness in `notdeft-xapian-ido-search-find-file'.

Additionally add `notdeft-mode-open-search' as an
`notdeft-open-search-function' implementation, now also taking plist
arguments.