remove `notdeft-select-note-file-by-search'
That customization option is somewhat redundant, since the
`notdeft-select-file-function' can be specified to be something that
uses search for selection or not.

This is nonetheless a breaking change in cases where the option has been
customized.

Make other minor simplifications and related renamings, for example to
shorten `notdeft-select-note-file' to `notdeft-select-file'.

Also, rename the original `notdeft-open-query' command more
descriptively as `notdeft-mode-open-query'.