update readme
Describe manual installation differently, in a more Emacs focused way,
instead of suggesting running makefiles, which is not as portable when
we know that Emacs must be available, but Make may not be.

Assume now that the notdeft-xapian executable is built from within
Emacs, and also that any rebuilds should happen automatically on demand.

Describe the setup steps in small steps in terms of Org code blocks,
which could potentially be evaluated from within the readme file.