add `notdeft-org-store-deft-link'
Use it to define a :store function for "deft:" links.

Notable here is that the implementation refuses to do much without the
`org-link-precise-link-target' function, which is not available in older
Org versions. When it is unavailable one can still execute
`notdeft-org-store-note-deft-link' directly to store a "deft:" link to
the note, without a more precise target.