don't rely on text properties in `notdeft-default-compread-file'
That is, don't rely on text properties being preserved when using
`completing-read', as in that general case we don't know which
implementation is being used, or whether it is one that does in fact
preserve them.

It looks like some do not, which would cause
`notdeft-default-compread-file' to return a nil file path regardless
of selection.